feat: rename ll_service to LLM_chatbot
